Separate and test normal estimation.
